+++ scripts/add_electricity.py
@@ -287,6 +287,10 @@ def load_costs(
     costs.at["solar", "capital_cost"] = costs.at["solar-utility", "capital_cost"]
     costs = costs.rename({"solar-utility single-axis tracking": "solar-hsat"})
 
+    # TODO Remove this temporary mapping once proper cost assumptions are implemented
+    costs.loc["solar-pv-utility"] = costs.loc["solar-utility"]
+    costs.loc["solar-pv-rooftop"] = costs.loc["solar-rooftop"]
+
     costs = costs.rename(columns={"standing losses": "standing_losses"})
 
     # Calculate storage costs if max_hours is provided
@@ -530,6 +534,7 @@ def attach_wind_and_solar(
     extendable_carriers: list | set,
     line_length_factor: float = 1.0,
     landfall_lengths: dict = None,
+    trajectories: pd.DataFrame = pd.DataFrame(),
 ) -> None:
     """
     Attach wind and solar generators to the network.
@@ -552,6 +557,9 @@ def attach_wind_and_solar(
         Factor to scale the line length, by default 1.0.
     landfall_lengths : dict, optional
         Dictionary containing the landfall lengths for offshore wind, by default None.
+    trajectories : pd.DataFrame, optional, by default None
+        DataFrame containing the trajectories for the current pyear to attach (p_nom_min and p_nom_max). When
+        provided, these values override any p_nom_max defined in the profile itself.
     """
     add_missing_carriers(n, carriers)
 
@@ -609,18 +617,37 @@ def attach_wind_and_solar(
             buses = ds.indexes["bus_bin"].get_level_values("bus")
             bus_bins = ds.indexes["bus_bin"].map(flatten)
 
-            p_nom_max = ds["p_nom_max"].to_pandas()
-            p_nom_max.index = p_nom_max.index.map(flatten)
-
-            p_max_pu = ds["profile"].to_pandas()
-            p_max_pu.columns = p_max_pu.columns.map(flatten)
-
             if not ppl.query("carrier == @car").empty:
                 caps = ppl.query("carrier == @car").groupby("bus").p_nom.sum()
-                caps = pd.Series(data=caps, index=ds.indexes["bus"]).fillna(0)
+                caps = caps.reindex(
+                    ds.indexes["bus_bin"].get_level_values("bus")
+                ).fillna(0)
+                caps.index = ds.indexes["bus_bin"]
             else:
                 caps = pd.Series(index=ds.indexes["bus"]).fillna(0)
             caps.index = caps.index.map(flatten)
+
+            if not trajectories.empty:
+                p_nom_max_min = (
+                    trajectories.query("carrier == @car")
+                    .groupby("bus")[["p_nom_max", "p_nom_min"]]
+                    .sum()
+                )
+                p_nom_max_min = p_nom_max_min.reindex(
+                    ds.indexes["bus_bin"].get_level_values("bus"), fill_value=0
+                ).set_axis(ds.indexes["bus_bin"].map(flatten), axis=0)
+                p_nom_max = p_nom_max_min["p_nom_max"]
+                p_nom_min = p_nom_max_min["p_nom_min"]
+            elif "p_nom_max" in ds.data_vars:
+                p_nom_max = ds["p_nom_max"].to_pandas()
+                p_nom_max.index = p_nom_max.index.map(flatten)
+                p_nom_min = caps
+            else:
+                p_nom_max = np.inf
+                p_nom_min = caps
+
+            p_max_pu = ds["profile"].to_pandas()
+            p_max_pu.columns = p_max_pu.columns.map(flatten)
 
             n.add(
                 "Generator",
@@ -629,7 +656,7 @@ def attach_wind_and_solar(
                 bus=buses,
                 carrier=car,
                 p_nom=caps,
-                p_nom_min=caps,
+                p_nom_min=p_nom_min,
                 p_nom_extendable=car in extendable_carriers["Generator"],
                 p_nom_max=p_nom_max,
                 marginal_cost=costs.at[supcar, "marginal_cost"],