+++ doc/release_notes.rst
@@ -27,6 +27,8 @@ Upcoming Open-TYNDP Release
 
 * Introduce processing and preparation of PEMMDB v2.4 capacity, must-run, and availability data, along with expansion trajectories for conventional and renewable power generation, electrolysers, batteries, and DSR (https://github.com/open-energy-transition/open-tyndp/pull/97).
 
+* Attach both solar and onwind technologies using PEMMDB and PECD data (https://github.com/open-energy-transition/open-tyndp/pull/115). Offwind statistics are also improved to include H2 generator capacities (in `MW_e`).
+
 
 Upcoming PyPSA-Eur Release
 ================